refactor(core): share `double` border rail geometry; migrate docx
Extract the floored-thirds device-pixel `double`-border geometry into core
(`draw/double-border.ts`: `doubleRailGeometry` + a fill-based `fillDoubleBorder`
painter), alongside `crispOffset` (canvas/crisp.ts) and the dash arrays
(draw/dash.ts). docx's `drawBorderLine` now calls `fillDoubleBorder` instead of a
local copy; the geometry unit test moves from docx to core. Behaviour-preserving
— docx VRT 21/21 unchanged, the relocated render path is byte-identical.

xlsx was evaluated for the same migration and deliberately NOT migrated. Excel's
double (§18.18.3) is two symmetric 1px strokes at ±1px with corner extension/trim
so a double BOX closes cleanly; the core fill helper's floored-thirds band has a
+0.5-device centring bias and no corner handling. Trialling the swap measured
~0.3% further from the Excel references (sample-11) — a fidelity regression — so
the two stay distinct: different Word/Excel conventions, not duplicated math. A
comment in xlsx `renderBorder` records this so the migration isn't re-attempted.
